Improvement: Merge render right view for laser and cnc
