+++ src/app/ui/pages/CncLaserShared/JobTypeModal.jsx
